refactor: Externalize the realtime initialization
We now pass the `getApp` function in parameter (it was previously a closure) but there should be no change to the underlying function.
